+++ toolchains/persee/v100/environment.sh
@@ -14,7 +14,13 @@ else
     . /data/gyselarunner/spack-1.1.0/share/spack/setup-env.sh
 fi
 
+# The hdf5 package is injecting the environment view `lib` path to `LD_LIBRARY_PATH`
+# which causes spurious segfaults for system executables, we manually remove it.
+LD_LIBRARY_PATH_TMP="$LD_LIBRARY_PATH"
 spack env activate gyselalibxx-env-omp-cuda
+export LD_LIBRARY_PATH="$LD_LIBRARY_PATH_TMP"
+unset LD_LIBRARY_PATH_TMP
+
 export OMP_PROC_BIND=spread
 export OMP_PLACES=threads
 export OMP_NUM_THREADS=8